+++ packages/tasit-action/src/Contract.test.js
@@ -3,7 +3,7 @@ import Account from "tasit-account";
 
 // Note: SampleContract.json is originally generated by `tasit-contracts` and was pasted here manually
 // See https://github.com/tasitlabs/TasitSDK/issues/45
-import { abi as contractABI } from "./testHelpers/SampleContract.json";
+import { abi as contractABI } from "../../tasit-contracts/build/contracts/SampleContract.json";
 
 // Note: Under the current `tasit-contracts` setup SampleContract aways will deployed with this address
 // See https://github.com/tasitlabs/TasitSDK/issues/138
