+++ packages/tasit-contracts/package.json
@@ -28,7 +28,6 @@
   },
   "devDependencies": {
     "chai-as-promised": "^7.1.1",
-    "ethers": "^4.0.23",
     "npm-run-all": "^4.1.5",
     "truffle": "^5.0.2"
   },
